+++ include/El/core/DistMatrix/CMakeLists.txt
@@ -1,6 +1,6 @@
 # Add the headers for this directory
 set_full_path(THIS_DIR_HEADERS
-  Abstract.hpp
+  AbstractDistMatrix.hpp
   Block.hpp
   Element.hpp
   )
